TY-2979 move feed logic to rust (3): changeConfiguration (#524)
**References**

- [TY-2979]
- requires #519

**Summary**

- add `FeedConfig` and `SearchConfig` & add corresponding ffi in `InitConfig`
- align field access and function calls for `InitConfig` with its field declaration order for better readability
- add `Engine::configure()` & corresponding ffi
- remove `EventConfig` and use new logic in `SystemManager.changeConfiguration()` in dart
- update [confluence config docs](https://xainag.atlassian.net/wiki/spaces/M2D/pages/2476605445/Discovery+Engine+Configurations)


[TY-2979]: https://xainag.atlassian.net/browse/TY-2979?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

+++ discovery_engine_core/core/src/stack/ops/common.rs
@@ -19,10 +19,10 @@ type ItemsResult<I> = Result<Vec<I>, GenericError>;
 type Request<I> = JoinHandle<ItemsResult<I>>;
 
 pub(super) async fn request_min_new_items<I: Send>(
-    max_requests: u32,
+    max_requests: usize,
     min_articles: usize,
     page_size: usize,
-    request_fn: impl Fn(u32) -> Request<I> + Send + Sync,
+    request_fn: impl Fn(usize) -> Request<I> + Send + Sync,
     filter_fn: impl Fn(Vec<I>) -> ItemsResult<I> + Send + Sync,
 ) -> ItemsResult<I> {
     let mut items = Vec::with_capacity(min_articles);
@@ -66,10 +66,10 @@ pub(super) async fn request_min_new_items<I: Send>(
 mod tests {
     use super::*;
 
-    struct Response(ItemsResult<u32>);
+    struct Response(ItemsResult<usize>);
 
     impl Response {
-        fn ok(items: &[u32]) -> Self {
+        fn ok(items: &[usize]) -> Self {
             Self(Ok(items.to_owned()))
         }
 
@@ -77,7 +77,7 @@ mod tests {
             Self(Err(GenericError::from(msg)))
         }
 
-        fn request(self) -> Request<u32> {
+        fn request(self) -> Request<usize> {
             tokio::spawn(async { self.0 })
         }
     }
